+++ instrumentation/opentelemetry-instrumentation-tornado/tests/test_instrumentation.py
@@ -157,8 +157,6 @@ class TestTornadoInstrumentation(TornadoTest):
         mock_span = Mock()
         mock_span.is_recording.return_value = False
         mock_tracer.start_span.return_value = mock_span
-        mock_tracer.use_span.return_value.__enter__ = mock_span
-        mock_tracer.use_span.return_value.__exit__ = True
         with patch("opentelemetry.trace.get_tracer") as tracer:
             tracer.return_value = mock_tracer
             self.fetch("/")